+++ app/src/main/java/com/example/ezvault/utils/ItemListFilter.java
@@ -3,8 +3,6 @@ package com.example.ezvault.utils;
 import androidx.annotation.NonNull;
 
 import com.example.ezvault.model.Item;
-import com.example.ezvault.model.utils.ItemListView;
-import com.example.ezvault.utils.IItemFilter;
 
 import java.util.ArrayList;
 import java.util.Iterator;
