+++ app/src/main/java/com/example/ezvault/utils/SerialPrediction.java
@@ -1,4 +1,4 @@
-package com.example.ezvault.model;
+package com.example.ezvault.utils;
 
 import androidx.annotation.NonNull;
