+++ app/src/main/java/com/example/ezvault/view/fragment/AddItemFragment.java
@@ -44,13 +44,13 @@ import com.example.ezvault.database.ItemDAO;
 import com.example.ezvault.database.RawUserDAO;
 import com.example.ezvault.model.Image;
 import com.example.ezvault.model.Item;
-import com.example.ezvault.model.ItemBuilder;
+import com.example.ezvault.utils.ItemBuilder;
 import com.example.ezvault.model.ItemList;
-import com.example.ezvault.model.SerialPrediction;
-import com.example.ezvault.model.SerialPredictor;
+import com.example.ezvault.utils.SerialPrediction;
+import com.example.ezvault.utils.SerialPredictor;
 import com.example.ezvault.model.Tag;
 import com.example.ezvault.model.User;
-import com.example.ezvault.textwatchers.NonEmptyTextWatcher;
+import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.upcAPI;
 import com.example.ezvault.utils.FileUtils;
 import com.example.ezvault.utils.TaskUtils;
