+++ app/src/main/java/com/example/ezvault/view/fragment/EditItemDetails.java
@@ -47,9 +47,9 @@ import com.example.ezvault.database.ImageDAO;
 import com.example.ezvault.database.ItemDAO;
 import com.example.ezvault.model.Image;
 import com.example.ezvault.model.Item;
-import com.example.ezvault.model.SerialPrediction;
-import com.example.ezvault.model.SerialPredictor;
-import com.example.ezvault.textwatchers.NonEmptyTextWatcher;
+import com.example.ezvault.utils.SerialPrediction;
+import com.example.ezvault.utils.SerialPredictor;
+import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.upcAPI;
 import com.example.ezvault.utils.FileUtils;
 import com.example.ezvault.utils.TaskUtils;
