+++ app/src/main/java/com/example/ezvault/view/fragment/LoginFragment.java
@@ -19,7 +19,7 @@ import com.example.ezvault.R;
 import com.example.ezvault.authentication.authentication.AuthenticationHandler;
 import com.example.ezvault.authentication.authentication.EmailPasswordAuthenticationStrategy;
 import com.example.ezvault.database.FirebaseBundle;
-import com.example.ezvault.textwatchers.NonEmptyTextWatcher;
+import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.utils.FragmentUtils;
 import com.example.ezvault.utils.UserManager;
