+++ app/src/main/java/com/example/ezvault/view/fragment/ProfileFragment.java
@@ -19,8 +19,8 @@ import com.example.ezvault.database.FirebaseBundle;
 import com.example.ezvault.database.RawUserDAO;
 import com.example.ezvault.database.UserService;
 import com.example.ezvault.model.ItemList;
-import com.example.ezvault.textwatchers.NonEmptyTextWatcher;
-import com.example.ezvault.textwatchers.PasswordWatcher;
+import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
+import com.example.ezvault.utils.textwatchers.PasswordWatcher;
 import com.example.ezvault.utils.UserManager;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.Tasks;
